requirements.txt
# Octave Clone MVP - Dependencies

# Core Framework
agno>=0.11.0

# Web Scraping
firecrawl-py>=1.0.0

# AI/LLM
openai>=1.0.0
anthropic>=0.40.0

# Data Models & Validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
